attempt to initiate a 3D frontend
