+++ kymatio/scattering3d/tests/test_torch_scattering3d.py
